requirements.txt
aioredis==2.0.1
anyio==4.3.0
apitally==0.11.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
certifi==2024.2.2
click==8.1.7
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
ollama==0.2.1
orjson==3.10.0
pdfme==0.4.11
pillow==10.4.0
pycountry==23.12.11
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
schedule==1.2.2
schematics==2.1.1
sniffio==1.3.1
starlette==0.37.2
starlette-decoRouter==1.0.0
Starlette-Login==0.2.1
starlette_htmx==0.1.1
StringGenerator==0.4.4
typing_extensions==4.10.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==2.3.6
